refactor(LottoNubmer): 숫자 형식이 아닐 경우 IllegarlArgumentException에서 NumberFormatException으로 변경, 패키지 구조 변경
