+++ src/main/java/lotto/controller/LottoController.java
@@ -3,10 +3,10 @@ package lotto.controller;
 import java.io.IOException;
 import java.util.List;
 import lotto.model.LottoRank;
-import lotto.model.LottoTicket;
+import lotto.model.lottonumbers.LottoTicket;
 import lotto.model.LottoTicketFactory;
 import lotto.model.Money;
-import lotto.model.WinningNumbers;
+import lotto.model.lottonumbers.WinningNumbers;
 import lotto.model.WinningStatistics;
 import lotto.view.InputView;
 import lotto.view.OutputView;
