+++ src/main/java/lotto/model/lottonumbers/LottoTicket.java
@@ -1,7 +1,8 @@
-package lotto.model;
+package lotto.model.lottonumbers;
 
 import java.util.Collections;
 import java.util.List;
+import lotto.model.LottoNumber;
 
 public class LottoTicket extends LottoNumbers {
     public LottoTicket(List<LottoNumber> lottoNumbers) {
